+++ airobas/blocks_hub/mip_blocks_lib/commons/bounds_propagation/bounds_computation_interface.py
@@ -1,11 +1,10 @@
 from abc import abstractmethod
 
-from blocks_hub.mip_blocks_lib.commons.bounds_propagation.linear_propagation import (
+from airobas.blocks_hub.mip_blocks_lib.commons.bounds_propagation.linear_propagation import (
     compute_bounds_linear_propagation,
 )
-from blocks_hub.mip_blocks_lib.commons.neural_network import NeuralNetwork
-from blocks_hub.mip_blocks_lib.commons.parameters import Bounds
-
+from airobas.blocks_hub.mip_blocks_lib.commons.neural_network import NeuralNetwork
+from airobas.blocks_hub.mip_blocks_lib.commons.parameters import Bounds
 
 def default_propagation(neural_network: NeuralNetwork):
     compute_bounds_linear_propagation(
