+++ airobas/blocks_hub/mip_blocks_lib/commons/bounds_propagation/bounds_computation_utils.py
@@ -2,17 +2,13 @@ try:
     from decomon.models.utils import ConvertMethod
 except ImportError as e:
     print(e)
-from blocks_hub.mip_blocks_lib.commons.bounds_propagation.bounds_computation_decomon import (
+from airobas.blocks_hub.mip_blocks_lib.commons.bounds_propagation.bounds_computation_decomon import (
     BoundsComputationDecomon,
 )
-from blocks_hub.mip_blocks_lib.commons.bounds_propagation.bounds_computation_interface import (
-    BoundComputation,
-)
-from blocks_hub.mip_blocks_lib.commons.bounds_propagation.bounds_computation_symbolic_arithmetic import (
+from airobas.blocks_hub.mip_blocks_lib.commons.bounds_propagation.bounds_computation_symbolic_arithmetic import (
     BoundComputationSymbolicArithmetic,
-    default_propagation,
 )
-from blocks_hub.mip_blocks_lib.commons.parameters import (
+from airobas.blocks_hub.mip_blocks_lib.commons.parameters import (
     ParamsBoundComputation,
     ParamsBoundComputationEnum,
 )
