+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools>=64", "setuptools_scm"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "AIROBAS"
+name = "airobas"
 description = "AI ROBustness ASsessment"
 readme = "README.md"
 authors = [
@@ -37,6 +37,7 @@ dev = ["tox>=4.6.4"]
 
 [tool.setuptools.packages.find]
 where = ["."]
+include = ["airobas*"]
 
 [tool.setuptools_scm]
