+++ tests/test_imports.py
@@ -2,8 +2,12 @@ import pytest
 
 
 def test_imports():
-    import gurobipy
     import gurobi_ml
     import maraboupy
+    from airobas.blocks_hub.decomon_block import DecomonBlock
+    from airobas.blocks_hub.marabou_block import MarabouBlock
+    from airobas.blocks_hub.gml_mip_block import GMLBrick
+    import keras
+    import decomon
     # import verif_pipeline
-    from blocks_hub.adv_block import CleverhansAdvBlock, CleverHansMultiIndexAdvBlock
+
